+++ Scripts/update_build_number.sh
@@ -1,11 +1,16 @@
 #!/usr/bin/env bash
 
-git rev-parse 2> /dev/null > /dev/null || { echo >&2 "Git repository required. Aborting build number update script."; exit 0; }
+set -euo pipefail
+
+if ! git rev-parse 2> /dev/null > /dev/null; then
+  echo >&2 "error: Git repository required. Aborting build number update script."
+  exit 1
+fi
 
 # Create version
-commitCount=`git log --oneline | wc -l | awk '{gsub(/^ +| +$/,"")} {print $0}'`
-commitTimestamp=`git log -1 --format=%ct | awk '{gsub(/^ +| +$/,"")} {print $0}'`
-dateRelativeToCommit=$((((`date "+%s"` - $commitTimestamp) / (15 * 60)) + 1))
+commitCount=$(git log --oneline | wc -l | awk '{gsub(/^ +| +$/,"")} {print $0}')
+commitTimestamp=$(git log -1 --format=%ct | awk '{gsub(/^ +| +$/,"")} {print $0}')
+dateRelativeToCommit=$(((($(date "+%s") - commitTimestamp) / (15 * 60)) + 1))
 version="$commitCount.$dateRelativeToCommit"
 echo "Build number is $version"
 
@@ -15,26 +20,32 @@ echo "Updating '$filepath'"
 /usr/libexec/PlistBuddy -c "Set :CFBundleVersion ${version}" "${filepath}"
 
 # Only update version numbers for extensions during archive
-if [ $ACTION = "install" ]; then
+if [ "$ACTION" = "install" ]; then
+  APP_PATH="${BUILT_PRODUCTS_DIR}/${CONTENTS_FOLDER_PATH}"
 
-	# Extensions info.plist
-	echo "Searching in '${BUILT_PRODUCTS_DIR}/${CONTENTS_FOLDER_PATH}/*.app/Info.plist'"
-	for subplist in "${BUILT_PRODUCTS_DIR}/${CONTENTS_FOLDER_PATH}"/*.app/Info.plist; do
-		echo "Updating $subplist"
-		/usr/libexec/PlistBuddy -c "Set :CFBundleVersion ${version}" "${subplist}"
-	done
+  # Extensions info.plist
+  echo "Searching in '${APP_PATH}/*.app/Info.plist'"
+  find "$APP_PATH" -name '*.app' -type d | while read -r subapp; do
+    if [ -f "$subapp/Info.plist" ]; then
+      echo "Updating $subapp/Info.plist"
+      /usr/libexec/PlistBuddy -c "Set :CFBundleVersion ${version}" "${subapp}/Info.plist"
+    fi
+  done
 
-	echo "Searching in '${BUILT_PRODUCTS_DIR}/${CONTENTS_FOLDER_PATH}/PlugIns/*.appex/Info.plist'"
-	for subplist in "${BUILT_PRODUCTS_DIR}/${CONTENTS_FOLDER_PATH}"/PlugIns/*.appex/Info.plist; do
-		echo "Updating $subplist"
-		/usr/libexec/PlistBuddy -c "Set :CFBundleVersion ${version}" "${subplist}"
-	done
-
+  echo "Searching in '${APP_PATH}/PlugIns/*.appex/Info.plist'"
+  if [ -d "$APP_PATH/PlugIns" ]; then
+    find "$APP_PATH/PlugIns" -name '*.appex' -type d | while read -r subapp; do
+      if [ -f "$subapp/Info.plist" ]; then
+        echo "Updating $subapp/Info.plist"
+        /usr/libexec/PlistBuddy -c "Set :CFBundleVersion ${version}" "${subapp}/Info.plist"
+      fi
+    done
+  fi
 fi
 
 # Dsym info.plist
 filepath="${DWARF_DSYM_FOLDER_PATH}/${DWARF_DSYM_FILE_NAME}/Contents/Info.plist"
 if [ -f "$filepath" ]; then
-	echo "Updating dSYM at '${filepath}'"
-	/usr/libexec/PlistBuddy -c "Set :CFBundleVersion ${version}" "${filepath}" || true
+  echo "Updating dSYM at '${filepath}'"
+  /usr/libexec/PlistBuddy -c "Set :CFBundleVersion ${version}" "${filepath}" || true
 fi